Master of Things Platform (MOT)/smart gate/requirements.txt
# python3 -m venv .venv
# .venv\Scripts\activate.bat
# pip install -r requirements.txt

Werkzeug==2.2.2
paho-mqtt


